# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Gym Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sets/{id}:
    put:
      operationId: updateSet
      tags: [sets]
      summary: Update a set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSetBody"
      responses:
        "200":
          description: Updated set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSet
      tags: [sets]
      summary: Delete a set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    SetRecord:
      type: object
      properties:
        id:
          type: string
        sessionExerciseId:
          type: string
        setNumber:
          type: integer
        reps:
          type: integer
          nullable: true
        weightKg:
          type: number
          nullable: true
      required:
        - id
        - sessionExerciseId
        - setNumber
...
    UpdateSetBody:
      type: object
      properties:
        reps:
          type: integer
          nullable: true
        weightKg:
          type: number
          nullable: true
        setNumber:
          type: integer
          nullable: true